+++ app/code/Magento/MediaContent/etc/di.xml
@@ -16,7 +16,7 @@
     <preference for="Magento\MediaContentApi\Api\Data\ContentIdentityInterface" type="Magento\MediaContent\Model\ContentIdentity"/>
     <preference for="Magento\MediaContentApi\Api\Data\ContentAssetLinkInterface" type="Magento\MediaContent\Model\ContentAssetLink"/>
     <preference for="Magento\MediaContentApi\Model\SearchPatternConfigInterface" type="Magento\MediaContent\Model\Content\SearchPatternConfig"/>
-    <preference for="Magento\MediaContentApi\Api\GetAssetIdByContentFieldInterface" type="Magento\MediaContent\Model\GetAssetIdByContentFieldComposite"/>
+    <preference for="Magento\MediaContentApi\Api\GetAssetIdsByContentFieldInterface" type="Magento\MediaContentApi\Model\Composite\GetAssetIdsByContentField"/>
     <type name="Magento\MediaGalleryApi\Api\DeleteAssetsByPathsInterface">
         <plugin name="remove_media_content_after_asset_is_removed_by_path" type="Magento\MediaContent\Plugin\MediaGalleryAssetDeleteByPath" />
     </type>
